Support xfail test variants.
Allow passing an extra `xfail` argument to the compile form
of the test macro to verify we reject certain inputs.

This is a little complicated. We can't just insert the
extra #[should_panic] in the xfail case before the expansion
of inner_cheddar_cmp_test!() because attributes do not
attach properly through nested macro expansions. Instead
we must match on the attributes in the inner pattern and
pass them explicitly to each invocation.